this connector should not change the key of the cacheEvent. Instead, the DD4T DefaultCacheAgent should do that.
