+++ src/main/java/org/dd4t/cache/CacheEventSerializer.java
@@ -1,30 +1,30 @@
 package org.dd4t.cache;
 
-import com.fasterxml.jackson.core.JsonProcessingException;
-import com.fasterxml.jackson.databind.ObjectMapper;
-import com.fasterxml.jackson.databind.module.SimpleModule;
+import com.fasterxml.jackson.core.JsonGenerationException;
+import com.fasterxml.jackson.core.JsonGenerator;
+import com.fasterxml.jackson.databind.SerializerProvider;
+import com.fasterxml.jackson.databind.ser.std.StdSerializer;
 import com.tridion.cache.CacheEvent;
 
 import java.io.IOException;
 
-public class CacheEventSerializer {
-    private static final ObjectMapper OBJECT_MAPPER = new ObjectMapper();
+public class CacheEventSerializer extends StdSerializer<CacheEvent> {
 
-    static {
-        final SimpleModule module = new SimpleModule();
-        module.addSerializer(CacheEvent.class, new CacheEventCustomSerializer());
-        module.addDeserializer(CacheEvent.class, new CacheEventDeserializer());
-        OBJECT_MAPPER.registerModule(module);
+    public CacheEventSerializer() {
+        this(null);
     }
 
-    private CacheEventSerializer() {
-
+    public CacheEventSerializer(Class<CacheEvent> vc) {
+        super(vc);
     }
 
-    public static String serialize(final CacheEvent cacheEvent) throws JsonProcessingException {
-        return OBJECT_MAPPER.writeValueAsString(cacheEvent);
-    }
-    public static CacheEvent deserialize(final String s) throws IOException {
-        return OBJECT_MAPPER.readValue(s, CacheEvent.class);
+
+    @Override
+    public void serialize(CacheEvent cacheEvent, JsonGenerator jsonGenerator, SerializerProvider serializerProvider) throws IOException, JsonGenerationException {
+        jsonGenerator.writeStartObject();
+        jsonGenerator.writeNumberField("type", cacheEvent.getEventType());
+        jsonGenerator.writeStringField("regionPath", cacheEvent.getRegionPath());
+        jsonGenerator.writeStringField("key", (String) cacheEvent.getKey());
+        jsonGenerator.writeEndObject();
     }
 }
